requirements.txt
tronpy==0.6.2
mnemonic==0.19
python-dotenv==1.0.0